Nasal: Fix double to degree conversion functions
Sign error: convertDoubleToDegree used to store the sign in the degree
field. For values between -1 and 0, this gives a value of -0 and the
sign is lost. Moved the sign to a separate field instead.

Improved rounding: convertDoubleToDegree used to sometimes return a
value ending with 60''. This should not be the case anymore.