Model: Make Rb05 flare visible from far away
Somewhat concerned about the way this is done:
The (ALS procedural) light is scaled with a 'dist-scale' to remain
visible. However, with the previous version it disappeared around 2km
depending on zoom.
By adding a regular 'scale' animation (factor x10), it remains visible
much further away. However adding that factor in the 'dist-scale' does
not work.

I have one hypothesis: FG computes a maximum render distance for objects
based on their size, and takes 'scale' into account in it, but not
'dist-scale'.