Autopilot: Fix warning lights
Implemented logic
Autopilot downgrade triggers the corresponding warning light, blinking.
Master warning acknowledgement converts it to steady light.
After this, pressing any of the 3 autopilot push-buttons resets the
warning light.

Precise documentation of the autopilot warning lights is hard to find.
This behaviour is described in the AJS SFI part 1, but it is not clear
if it only applies to certain types of warnings, or all of them.