RCS: Fix test when a target gets in range
The RCS test has a caching mechanism to avoid redoing the tests all the
time. However, that cache was discarded based on the number of queries
done since the last update, which means it could give completely wrong
answers for a target which had not been queried for a long time.
This causes targets outside RCS range to sometimes be displayed when
they first come into range.
Fixed by setting a maximum validity time for each cached result.