Sim: Fixes in FG/Viggen flightplans interface
Rework the custom Viggen flight plan delegate (flightPlanDelegate.nas).
Re-enable the default route_manager.RouteManagerDelegate,
which was essentially duplicated in the Viggen code.
Disable the default route_manager.DefaultGPSDelegate,
and replace it to implement a customized waypoints sequencing.

The custom sequencing code allows to keep the flightplan active when
reaching the last waypoint, and thus to remove the existing
're-activate flightplan when finished' code.
The latter was prone to infinite loops at the slightest change in
flightplan code, and it is a miracle that such loops did not occur
previously.

Fixed a re-entrance bug in route.Polygon._wpChanged, which caused a C++
assertion failure in FlightPlan::lockDelegates().
Adding a 0 timer to break any recursion is sufficient, if not elegant.