FDM: Rework landing gear steering and brakes
Remove the ability to disable nose wheel steering (there is no reference
to such thing in the flight manual).

Nose wheel steering limit is 30° no matter the rolling speed.
However steering is disabled when the nose gear is not compressed.

The previous two changes make it harder to control the direction during
landing/takeoff.
Implement a deadband of up to 4°, described in the manual, to help with
steering at high speed (it acts as a damper for steering).

Remove nose wheel brakes.
Slightly increase friction coefficient to compensate.