HUD: Shader for HUD parallax correction
Add parallax correction to the HUD shader. It keeps the HUD texture
(canvas) straight in front of the view point at all time.

Compared to the previous canvas-offsets-in-nasal method, it avoids any
lag between nasal computing the correction, and it being displayed.

It is not enabled, because the current HUD needs to be adapted first.