Arms: More realistic Rb 04 and Rb 15
Anti-ship missiles are fired without lock.
Change the range/field of view of their radars to the real values.
This is realistic for the Rb 04.
The Rb 15 would need some advanced waypoint navigation in addition...

Improve the engine power for the Rb 15, but significantly decrease its
range to 70km (I suspect older values were for more modern,
surface-to-surface versions of the Rb 15).