Remove combat mode. Allow firing weapons in NAV
For the JA, the existing 'combat mode' is actually the 'aiming mode'.
While I have no reference on this, I think that aiming mode is not
required to fire weapons (based on comparison with AJS, which can fire
most weapons in NAV/SPA modes...).

This commit removes 'combat mode', the master arm switch controlling it
(which really is used for training modes/weapon simulation), and
disables the HUD aiming mode.
Furthermore, it allows to fire weapons in any mode (TODO: check firing
restrictions).

I plan to re-implement aiming mode logic from scratch, making it similar
to the AJS combat mode: it makes the HUD focused on aiming, but it is
still possible (and practical) to use weapons in NAV mode.