JA HUD: COMBAT mode is dead, long live AIMING mode
As previously, 'h' toggles AIMING mode.
However it does not correspond to a physical switch.
(I think it should be the 'Optical aiming mode' on the stick,
unless that is yet another thing)

Due to previous commits, AIMING mode is not required to use weapons.
It is a decluttered mode intended for combat.
An advanced gun sight will be added.