Sound: Cleanup 'avionics' tags
Only use <type>avionics</type> for sounds which are supposed to be
heared from the pilot earphones.  Adapt some other sound parameters
(position, reference-dist) accordingly.

This ensures that the 'avionics' sound volume/settings have the expected
effect.
It also fixes that cockpit engine sounds were not working when switching
to an external view, then back to the cockpit view.