Autothrottle: Use g-load corrected alpha
Autothrottle in alpha hold mode doesn't actually maintain alpha:
alpha is corrected for g-load so as to maintain airspeed corresponding
to the target alpha at g-load=1.

This fixes the autothrottle wind-up tendency
(a pitch up moment causes an alpha increase, causing the autothrottle to
increase power, which increases the pitch up moment).