View: Fix helicopter view position for MP Viggens
Fixes a bug introduced by c8fbf665
When switching view to a MP Viggen with helicopter view, the camera
would start inside the aircraft due to a wrong z-offset

This was somehow caused by an alias in the view configuration.
I am still not completely sure why.
Replacing the alias with a constant fixes the issue, and is not a big deal.