GUI: Rework Fuel and Loadout dialog
Add payload preview using canvas.
Add payload selection drop down menus, similar to the Fuel and Equipment
dialog.

For the AJS only, rb24j/rb74 are no longer included in the main presets.
Instead the main preset only consists of A/G weapons, and sidewinders
are available as options, which modify the preset.  This allows more
choices, while reducing the length of the preset loadouts list.