GUI: Rework AJS loadout Rb 24J/74 options
Transform them in buttons to add IR missiles to the current loadout.
This makes the loadout part of the dialog stateless (which is nice),
and allows using them with hand-made loadouts.

For the outer wing pylons: two buttons to load/unload RB 24J.
These pylons are left unmodified when selecting a loadout.
(but still affected by 'Remove payload' and 'Reload').

For the main pylons: two buttons to add 1/2 RB 74.
The code to add them now works well on custom loadouts.
However, these will be removed when selecting a new loadout preset.