Fire control: Rework missile seekers logic
Seekers for A/A missiles (JA and AJS IR) are turned on as soon as the
missile is selected (instead of after unsafing).
For the AJS, this is confirmed by the AJS SFI part 2 (IR missile tone
can be obtained on the ground as long as the weapon selector knob is in
position IR-RB).

For the Rb-75, the seeker is turned on together with the HUD aiming
mode, i.e. in mode ANF or at unsafing, if firing circuits are enabled.

Also fix IR seeker bore position (should have been -0.8, but was
probably not applied in the AJS).