HUD: Efficient colour changes using cascading
The HUD used setColor() on the root group to change brightness.
This goes through all canvas elements recursively to set their colour
individually, which is horribly slow (purely due to properties i/o).

Use canvas cascading instead: the styles "fill" and "stroke" are set on
the root group only, and the colour propagates to all elements.

Problem is, we want "fill" only for text, and "stroke" only for paths,
and there is no way to restrict the styles set on the root (not CSS...).
As a workaround, all paths have "fill": "none" set, and text has
"stroke": "none", which overrides the global setting.

This is a bit silly, but the point is that in order to change the colour,
only the root style needs to be modified.

This is a huge performance improvement: with this one could reasonably
change the HUD colour at every HUD update loop (which will be important
to implement the HUD brightness adapting to ambient light).

There is even a noticeable performance difference simply when turning
the brightness knob.