Add JA37 elevation profile generation program
The goal is to simulate the JA37 elevation map of Sweden
(but for the entire world). With 'squares' of 6 arc minutes (~11km),
and 1 byte per square, this is only a few MB of data for the entire world.
Thus it is reasonable to precompute it all, and store it in a single file
(might split later if it is problematic for the Nasal side).

This commit adds a description of the file format used for this
elevation map, and a program to generate it from STRMGL30 DEM.
Said program simply averages elevations for each squares,
which ought to be improved.

Final elevation map not included in this commit.
It is automatically generated, and not exactly small, so I don't want
to clutter the git repo with several different versions of it.
I will only add it once I'm happy with the generation program.