Scheduler: Pass simulated time elapsed to loops
The scheduler update rate is tied to real time / frame rate.
But it also passes an optional 'dt' argument to some loops.
This was meant to represent elapsed simulated time, but was real time instead.

Also ensure that a ridiculous 'dt' is not passed when restarting
simulation, by checking if simulation time decrease.

The only affected systems were oxygen, and the environment.