Scheduler: Remove 'locals' loop parameter
I completely misunderstood the 'local' argument of call().
I thought it was a hash of additional variables the function may refer to,
whereas it is _the_ local environment of the function (arguments and
local variables are stored in it).

That led to a nasty bug with the experimental CI.
Thankfully it wasn't used otherwise.