AJS radar: Add terrain noise
Add noise to water to simulate waves, with strength depending on wind.
Add time-independent noise to terrain, with strength depending on bumpiness.
Add altitude-noise to urban areas, to simulate buildings.

Add noise and radar signal falloff from rain (not working very well).

Not picked in main branch because it is quite slow.
(~28fps -> ~25fps due to noise, in a not very formal test)