JA HUD: Dynamic distance scale for missiles
With selected missile and target, the HUD distance scale goes up to the
maximum radar range if the target is beyond maximum firing range.
(otherwise, the scale goes up to the maximum firing range as before)

Based on some HUD recording, this seems to be the correct behaviour,
at least for the Rb 74.