TI: Radar conv.: TI track selection
Fix selection of TI track (radar or DL), which can be used instead
of waypoints, and port it to new radar system.

The old system was partially broken because I maintained it poorly,
and the keybinding was missing since some time.
For the new system, the destination is not necessarily the primary target,
but any radar or datalink track. It is selected using the TI cursor.