AJS Radar: Use shaders for radar picture rendering
This is a proof of concept for shader radar rendering.
The goal is to move a large part of the radar rendering code from
canvas+nasal to GL shaders.

The main change is that nasal code draws radar image as a B-scope,
and at exactly the resolution used by the radar.
The shader does the coordinate transformation to draw it as a PPI.
For a given quality, this should minimize the number setPixel calls.

Using shaders is also the only reasonable way I can think of
to eventually implement radar image decay, which is very important for
the memory mode (and looks nice).

The canvas uses the three color channels independently:
- red: symbols (artificial horizon, navigation), rendered directly by the shader.
- green: radar picture (as B-scope)
- blue: age of radar picture (will be used for image decay)