Model: Redo interior/exterior split
Add new models {ja,aj}37-interior.ac for interior details which do not
fit elsewhere.
Delete the sidepanels.ac model, splitting it between interior and
exterior models.
Move details such as the HUD to the interior model.

The exterior model now contains the main cockpit sides, to properly
close it off. With the details removed, it is slightly lighter than
previously (~100KB), and uses fewer different effects.

The bottom and rear walls of the cockpit have been slightly modified and
simplified.

Additional cleanup of the main model in this commit:
- fix swapped naming of internal/outer glass objects
- cleanup some materials
- simplify details of fuselage on the nose and sides.
- simplify beacon lights.