CI: Duplicate FG generic shaders
The CI rendering shader is plugged into FG standard shaders to provide
the screen color (unfortunately, this requires to duplicate the shader code).

This version is based on FG 2020.13 shaders, and is compatible
with FG 2020.4+ (but the CI will be missing shadows, etc.).

When shaders are disabled, the CI is displayed using canvas only.
In this case, only navigation symbols are shown, and not the radar.