CI: Implement range and azimuth lines
Re-do range and azimuth lines with shader rendering.

To indicate to the shader which lines need to be drawn (function of
the azimuth) the blue channel of the canvas is used:
the blue channel contains information which only depends on the azimuth
(not on the range). Thus, a lot of information can be passed to the
shader, as several horizontal strips in the blue channel.

Currently two such strips are used: time of writing (already before this
commit) and radar range (added by this commit).
Other symbols (cross, aiming lines) may later be added similarly.

This commit also does a lot of cleanup in the CI shader:
- replace constant variables by preprocessor constants
- remove the multiple sampling for decay