Nav: Start custom AJS navigation system
The goal is to stop using the default route manager,
which is a very poor fit for what the AJS needs.

This commit adds constants to reference the various AJS waypoints,
and some basic route manager functionalities.
This new route manager is not used yet, and there is no way to load a route.