AJS Nav: Use new system for waypoint instruments
Define new properties
/instrumentation/waypoint-indicator/true-bearing-deg
/instrumentation/waypoint-indicator/dist-km
They are used as input for the JSBSim part of the waypoint instruments.

For the AJS, these properties are the output of the custom nav system,
replacing /autopilot/route-manager/wp/{true-bearing-deg,dist-km}.
Thus the new nav system now controls the instruments.

For the JA, [...]/true-bearing-deg replaces [...]/ja-bearing-deg as
output of the destination switching code in displays/common.nas.
[...]/dist-km is unused. Logic is otherwise unchanged.