Sim: Rework afterburner controls
When changin afterburner zone, simulate a throttle deadzone,
the ignition delay, and the sudden increase in power.

Change engine <augmethod> property to 2: augmentation thrust is
controlled by setting throttle between 1.0 (=mil) and 2.0 (=max aug).
This is easier to manipulate.

Move a lot of engine related JSBSim stuff to a new file (jsb-engine.xml).
It was split between jsb-flight.xml and jsb-controls.xml.