AJS Nav: Better conversions to nasal ghosts
AJS waypoints/airbases now have an optional waypoint/airport ghost
attached to them, to make conversions back to nasal ghost (when saving
flightplans in particular) nicer.

This is getting a bit silly, but for the airbases at least I don't see
a way to avoid it while keeping the ability to place airbases anywhere.