Model: Adjust left panel, add missing switches
Add tank pump and engine deice switches on the right console.
Add manual ignition switch and restart button on the left console.
Move the indicator lights test button to the right console.

Increase size of gauges on the left console, and move them closer to
keep them visible (using the space where the test button was).

Adjust test for a few switches to make it more readable.