Sim: Limit fuel distributor based on acceleration
Disable distributor based on acceleration instead of attitude.

The fuel distributor is disabled during unbalanced flight.
The manual is ambiguous on whether the limit angles are in terms of
attitude, or of the acceleration vector, but the latter seems much more
sensible given the goal (avoiding overfill).

This should help avoid draining the collector tank during dogfights.