Radar: (Re)add lock tone
I forgot to port the radar lock sound during the radar rework.
It now triggers whenever the radar starts gets valid tracking info on
the contact. This coincides with the MI symbology.

Add a 'ps46.hasValidTrack()' method to test this condition,
moving it from the MI to the radar code.