Split loading rwy from route from landing mode
landing-mode.nas currently is in charge of loading the runway infos from
the route manager, and displays go through landing-mode.nas when they
need these infos. This is making it painful to rewrite the landing mode.

Duplicate this code to a new file Nasal/navigation/runway.nas,
and redirect anyone outside landing-mode.nas to it.